Added container exec policy to db

+++ backend/internal/api/models/containers.go
@@ -2,6 +2,7 @@ package models
 
 import (
 	"context"
+	"log"
 
 	database "github.com/AlvaroParker/web-console/internal/api"
 	"github.com/docker/docker/api/types/container"
@@ -9,12 +10,15 @@ import (
 )
 
 type Terminal struct {
-	Id          int    `json:"id"`
-	ContainerID string `json:"containerid"`
-	Email       string `json:"email"`
-	Image       string `json:"image"`
-	Tag         string `json:"tag"`
-	Name        string `json:"name"`
+	Id             int    `json:"id"`
+	ContainerID    string `json:"containerid"`
+	Email          string `json:"email"`
+	Image          string `json:"image"`
+	Tag            string `json:"tag"`
+	Name           string `json:"name"`
+	AutoRemove     bool   `json:"auto_remove"`
+	NetworkEnabled bool   `json:"network_enabled"`
+	Command        string `json:"command"`
 }
 
 type TerminalRes struct {
@@ -34,14 +38,16 @@ type Container struct {
 	Command        *string `json:"command"`
 }
 
-func ValidateContainer(email string, hash string) bool {
-	var terminal Terminal
-	query := database.DB.QueryRow("SELECT * FROM terminals WHERE email = $1 and containerid = $2", email, hash)
-	queryErr := query.Scan(&terminal.Id, &terminal.ContainerID, &terminal.Email, &terminal.Image, &terminal.Tag, &terminal.Name)
+func ValidateContainer(email string, hash string) *Container {
+	var container Container
+	query := database.DB.QueryRow("SELECT image, tag, name, auto_remove, network_enabled, command FROM terminals WHERE email = $1 and containerid = $2", email, hash)
+	queryErr := query.Scan(&container.Image, &container.Tag, &container.Name, &container.AutoRemove, &container.NetworkEnabled, &container.Command)
+
 	if queryErr != nil {
-		return false
+		log.Println("[models.ValidateContainer] Error while querying the database: ", queryErr)
+		return nil
 	}
-	return true
+	return &container
 }
 
 func GetTerminals(email string) ([]TerminalRes, error) {
@@ -66,11 +72,27 @@ func GetTerminals(email string) ([]TerminalRes, error) {
 
 // Add a container ID to the database
 func AddContainerDB(email string, containerID string, container Container) error {
-	_, err := database.DB.Exec("INSERT INTO terminals (containerid, email, image, tag, name) VALUES ($1, $2, $3, $4, $5)", containerID, email, container.Image, container.Tag, container.Name)
+	_, err := database.DB.Exec("INSERT INTO terminals (containerid, email, image, tag, name, auto_remove, network_enabled, command) VALUES ($1, $2, $3, $4, $5, $6, $7, $8)",
+		containerID, email, container.Image, container.Tag, container.Name, container.AutoRemove, container.NetworkEnabled, container.Command)
 	return err
 }
 
 func DeleteContainerDB(id string, email string) (bool, error) {
+	// Check if the container is running
+	cli, err := client.NewClientWithOpts(client.FromEnv)
+	// Delete using docker client
+	if err != nil {
+		return false, err
+	}
+	response, errInspect := cli.ContainerInspect(context.Background(), id)
+	if errInspect != nil {
+		return false, errInspect
+	}
+
+	if response.State.Running {
+		return false, nil
+	}
+
 	sqlRes, errDB := database.DB.Exec("DELETE FROM terminals WHERE containerid = $1 and email = $2", id, email)
 	if errDB != nil {
 		return false, errDB
@@ -78,12 +100,6 @@ func DeleteContainerDB(id string, email string) (bool, error) {
 	rowsAffected, _ := sqlRes.RowsAffected()
 	if rowsAffected == 0 {
 		return false, nil
-	}
-
-	// Delete using docker client
-	cli, err := client.NewClientWithOpts(client.FromEnv)
-	if err != nil {
-		return false, err
 	}
 
 	errRmDocker := cli.ContainerRemove(context.Background(), id, container.RemoveOptions{})
@@ -102,4 +118,15 @@ func CountContainers(email string) (int, error) {
 	}
 
 	return count, nil
+}
+
+func GetContainerInfo(id string, email string) (*TerminalRes, error) {
+	query := database.DB.QueryRow("SELECT containerid, image, tag, name FROM terminals WHERE containerid = $1 and email = $2", id, email)
+	var terminal TerminalRes
+	errDb := query.Scan(&terminal.ContainerID, &terminal.Image, &terminal.Tag, &terminal.Name)
+	if errDb != nil {
+		return nil, errDb
+	}
+
+	return &terminal, nil
 }
